Dreamweaver/requirements.txt
openai-whisper
langchain-core
langchain-google-genai
python-dotenv
diffusers
transformers
accelerate
torch
safetensors
notion-client